rag_agent/pyproject.toml
[project]
name = "sql-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "click>=8.1.8",
    "google-adk>=0.0.3",
    "google-genai>=1.9.0",
    "litellm>=1.67.6",
    "pinecone>=6.0.2",
    "pinecone-datasets>=1.0.2",
    "pymilvus>=2.5.8",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
    "sseclient-py>=1.8.0",
    "streamlit>=1.45.0",
    "transformers>=4.51.3",
]
